Move create page logic into own file
