+++ gatsby-node.js
@@ -4,9 +4,6 @@ const { initStitch } = require('./src/utils/setup/init-stich');
 const {
     postprocessDocument,
 } = require('./src/utils/setup/postprocess-document');
-const {
-    getRelatedPagesWithImages,
-} = require('./src/utils/setup/get-related-pages-with-images');
 const { saveAssetFiles } = require('./src/utils/setup/save-asset-files');
 const {
     validateEnvVariables,
@@ -14,15 +11,11 @@ const {
 const { onCreatePage } = require('./src/utils/setup/on-create-page');
 const { createTagPageType } = require('./src/utils/setup/create-tag-page-type');
 const { getMetadata } = require('./src/utils/get-metadata');
-const { getNestedValue } = require('./src/utils/get-nested-value');
-const { getPageSlug } = require('./src/utils/get-page-slug');
-const { getSeriesArticles } = require('./src/utils/get-series-articles');
-const { getTemplate } = require('./src/utils/get-template');
 const {
     DOCUMENTS_COLLECTION,
     METADATA_COLLECTION,
-    SNOOTY_STITCH_ID,
 } = require('./src/build-constants');
+const { createArticlePage } = require('./src/utils/setup/create-article-page');
 
 // Consolidated metadata object used to identify build and env variables
 const metadata = getMetadata();
@@ -31,7 +24,7 @@ const DB = metadata.database;
 const PAGE_ID_PREFIX = `${metadata.project}/${metadata.user}/${metadata.parserBranch}`;
 
 // different types of references
-const PAGES = [];
+const pages = [];
 
 // in-memory object with key/value = filename/document
 const slugContentMapping = {};
@@ -66,7 +59,7 @@ exports.sourceNodes = async () => {
             doc,
             PAGE_ID_PREFIX,
             assets,
-            PAGES,
+            pages,
             slugContentMapping
         );
     });
@@ -89,34 +82,14 @@ exports.createPages = async ({ actions }) => {
     learnFeaturedArticles = allSeries.learn;
     delete allSeries.home;
     delete allSeries.learn;
-    PAGES.forEach(page => {
-        const pageNodes = slugContentMapping[page];
-
-        if (pageNodes && Object.keys(pageNodes).length > 0) {
-            const template = getTemplate(
-                getNestedValue(['ast', 'options', 'template'], pageNodes)
-            );
-            const slug = getPageSlug(page);
-            if (pageNodes.query_fields) {
-                const relatedPages = getRelatedPagesWithImages(
-                    pageNodes,
-                    slugContentMapping
-                );
-                pageNodes['query_fields'].related = relatedPages;
-            }
-            const seriesArticles = getSeriesArticles(allSeries, slug);
-            createPage({
-                path: slug,
-                component: path.resolve(`./src/templates/${template}.js`),
-                context: {
-                    metadata,
-                    seriesArticles,
-                    slug,
-                    snootyStitchId: SNOOTY_STITCH_ID,
-                    __refDocMapping: pageNodes,
-                },
-            });
-        }
+    pages.forEach(page => {
+        createArticlePage(
+            page,
+            slugContentMapping,
+            allSeries,
+            metadata,
+            createPage
+        );
     });
 
     const tagTypes = ['author', 'languages', 'products', 'tags', 'type'];
